requirements.txt
streamlit
pandas
twilio
markdown-it-py
openpyxl
